+++ extensions/iceworks-doctor/web/src/pages/Dashboard/components/ScanCard/components/ScanFailedWrap/index.tsx
@@ -1,0 +1,19 @@
+import * as React from 'react';
+import { Message } from '@alifd/next';
+import styles from './index.module.scss';
+
+const ScanFailedWrap = () => {
+  return (
+    <div className={styles.scanFailedWrap}>
+      <Message title="Error" type="error">
+        {window.USE_EN ? 'Scan Failed, open ' : '扫描失败，请至 '}
+        <a href="https://github.com/ice-lab/iceworks/issues" target="_blank">
+          https://github.com/ice-lab/iceworks/issues
+        </a>
+        {window.USE_EN ? ' report your problem' : ' 反馈'}
+      </Message>
+    </div>
+  );
+};
+
+export default ScanFailedWrap;