+++ extensions/iceworks-time-steward/src/timer.ts
@@ -10,9 +10,9 @@ const recorder = new Recorder(name, version);
 export class Timer {
   private disposable: vscode.Disposable;
 
-  private lastFile: string = '';
+  private lastFile = '';
 
-  private lastHeartbeat: number = 0;
+  private lastHeartbeat = 0;
 
   private user: { empId: string; account: string };
 
@@ -81,7 +81,7 @@ export class Timer {
             const subTime = time - this.lastHeartbeat;
             const { account, empId } = this.user;
             if (this.lastHeartbeat !== 0) {
-              recorder.record({
+              recorder.recordPV({
                 module: 'main',
                 action: 'tracking',
                 data: {
@@ -118,7 +118,7 @@ export class Timer {
     if (workspaceFolder) {
       try {
         const packageJsonResultObj: IPackageJson = JSON.parse(
-          fs.readFileSync(`${workspaceFolder.uri.path}/package.json`, 'utf-8')
+          fs.readFileSync(`${workspaceFolder.uri.path}/package.json`, 'utf-8'),
         );
         return packageJsonResultObj.name;
       } catch (e) {