+++ packages/doctor/src/getEslintReports.ts
@@ -7,7 +7,7 @@ export default function getBestPracticesReports(
   warningWeight: number,
   errorWeight: number,
   files: IFileInfo[],
-  fix?: boolean
+  fix?: boolean,
 ): IEslintReports {
   let warningScore = 0;
   let errorScore = 0;
@@ -25,12 +25,26 @@ export default function getBestPracticesReports(
 
   files.forEach((file) => {
     aliEslintCliEngine.executeOnText(file.source, file.path).results.forEach((result) => {
+      // Remove Parsing error
+      result.messages = (result.messages || []).filter((message) => {
+        if (message.fatal && message.severity === 2 && message.message.startsWith('Parsing error:')) {
+          result.errorCount--;
+          return false;
+        }
+        return true;
+      });
+
       reports.push({
         ...result,
         filePath: file.path,
       });
     });
   });
+
+  if (fix) {
+    // output fixes to disk
+    CLIEngine.outputFixes(aliEslintCliEngine.executeOnFiles(files.map((file) => file.path)));
+  }
 
   reports.forEach((report) => {
     warningScore += report.warningCount * warningWeight;