+++ scripts/publish-beta-extension.ts
@@ -5,6 +5,8 @@ import * as path from 'path';
 import { spawnSync } from 'child_process';
 import uploadExtesions from './upload-extensions';
 import { IExtensionInfo, getExtensionInfos } from './getExtensionInfos';
+import sleep from './fn/sleep';
+import checkPackagePublished from './fn/checkPackagePublished';
 import extensionDepsInstall from './fn/extension-deps-install';
 import updateExtensionDependencies from './fn/updateExtensionDependencies';
 
@@ -17,38 +19,47 @@ function publish(extension: string, directory: string, version: string): void {
   });
 }
 
-// Entry
-console.log('[PUBLISH BETA] Start:');
-getExtensionInfos().then((extensionInfos: IExtensionInfo[]) => {
-  const shouldPublishExtensions: IExtensionInfo[] = [];
+async function start() {
+  try {
+    await checkPackagePublished(true);
+    await sleep(50000);
+    console.log('[PUBLISH BETA] Start:');
+    const extensionInfos: IExtensionInfo[] = await getExtensionInfos();
+    const shouldPublishExtensions: IExtensionInfo[] = [];
 
-  for (let i = 0; i < extensionInfos.length; i++) {
-    const { name, directory, shouldPublish } = extensionInfos[i];
-    if (shouldPublish) {
-      // Update extension package json
-      updateExtensionDependencies(name, directory);
-      // Update inside web project package json
-      updateExtensionDependencies(name, path.join(directory, 'web'));
+    for (let i = 0; i < extensionInfos.length; i++) {
+      const { name, directory, shouldPublish } = extensionInfos[i];
+      if (shouldPublish) {
+        // Update extension package json
+        updateExtensionDependencies(name, directory);
+        // Update inside web project package json
+        updateExtensionDependencies(name, path.join(directory, 'web'));
 
-      shouldPublishExtensions.push(extensionInfos[i]);
+        shouldPublishExtensions.push(extensionInfos[i]);
+      }
     }
+
+    // npm install
+    extensionDepsInstall();
+
+    // Publish
+    let publishedCount = 0;
+    const publishedExtensions = [];
+    for (let i = 0; i < shouldPublishExtensions.length; i++) {
+      const { name, directory, localVersion } = shouldPublishExtensions[i];
+      publishedCount++;
+      console.log(`--- ${name}@${localVersion} ---`);
+
+      publish(name, directory, localVersion);
+      publishedExtensions.push(`${name}:${localVersion}`);
+    }
+    uploadExtesions(publishedExtensions);
+    console.log(`[PUBLISH EXTENSION BETA] Complete (count=${publishedCount}):`);
+    console.log(`${publishedExtensions.join('\n')}`);
+  } catch (e) {
+    console.error(e);
+    process.exit(1);
   }
+}
 
-  // npm install
-  extensionDepsInstall();
-
-  // Publish
-  let publishedCount = 0;
-  const publishedExtensions = [];
-  for (let i = 0; i < shouldPublishExtensions.length; i++) {
-    const { name, directory, localVersion } = shouldPublishExtensions[i];
-    publishedCount++;
-    console.log(`--- ${name}@${localVersion} ---`);
-
-    publish(name, directory, localVersion);
-    publishedExtensions.push(`${name}:${localVersion}`);
-  }
-  uploadExtesions(publishedExtensions);
-  console.log(`[PUBLISH EXTENSION BETA] Complete (count=${publishedCount}):`);
-  console.log(`${publishedExtensions.join('\n')}`);
-});
+start();